Pack the per-cell, per-item and per-action structures
CellData drops its row/col (the caller knows the position: format(row,
col)) and keeps a u32 indent: 56 -> 32 bytes. ActionInst stores its anchor
as a packed u32 (ItemId::pack/unpack): 24 -> 16 bytes. RecEntry::Many points
into a side table of ranges: 24 -> 12 bytes. The item index and Range use
u32 indices, CellItem.index/span are u32 (88 -> 72 bytes).
CellPredicate::test takes the cell position explicitly.

stack_test11: peak RSS 395 -> 317 MB, wall 0.83 s -> 0.76 s; at k=4
1672 -> 1255 MB, 3.27 s -> 2.94 s.

+++ src/tests.rs
@@ -129,7 +129,7 @@ fn end_to_end_match_and_interpret() {
         .iter()
         .find(|it| it.row == 1 && it.col == 0)
         .unwrap();
-    assert_eq!(atomic.span, (0, "IKT".len()));
+    assert_eq!(atomic.span, (0, "IKT".len() as u32));
 
     let out = interpret(&InterpreterCfg::default(), &syntax, &sem, None).unwrap();
     assert!(out.diagnostics.is_empty());
